requirements.txt
fastapi[standard]
uvicorn[standard]
SQLAlchemy~=2.0.32
psycopg2-binary
alembic~=1.13.2
sqlalchemy[asyncio]
asyncpg
python-jose
passlib~=1.7.4
pydantic~=2.8.2
python-dotenv~=1.0.1
pydantic-settings